move DeleteAccountPopover into own component
